reorganizando paineis de acervo e corrigindo bug de menu acervo
