+++ src/main/java/view/acervo/PainelAcervoCadastro.java
@@ -1,4 +1,4 @@
-package view;
+package view.acervo;
 
 import java.awt.Color;
 import java.awt.GridLayout;
