+++ src/main/java/view/acervo/PainelAcervoEditar.java
@@ -1,4 +1,4 @@
-package view;
+package view.acervo;
 
 import javax.swing.JPanel;
 import javax.swing.JMenuBar;
